refactor: move format_age and worktree_modified_files to utils (#150)
